+++ src/test/java/seedu/address/model/types/person/PersonTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.types.person;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
